Organize folders and path to assets
