+++ src/assets/images/player.png
